Fix matching in non-convex contexts
Main issue is that matchWire / matchPort logic is mixed up.
Both need to know if the source port is a Boundary or not, but only because
matchPort isn't written generically for "Source" and "Target" ports.

As usual, this is fixable with lenses- just parametrise matchPort by the lens
corresponding to the field _matchStatePortsL / _matchStatePortsR