WIP: big refactor of rendering code
Main changes: now the Cartographer.Draw module handles rendering in "abstract
tile coordinates" - i.e., everything is given a position on an integer grid.

The idea is that consumers of the Draw module will simply iterate over each of
the lists to produce an image- i.e., rendering can be "compositional" and
render each element independently.