Fix issue #1
We now re-lay-out the graph after every call to rewriteLayout.
This isn't ideal, because it doesn't necessarily preserve the geometry of the
pattern - i.e., it might cause the substituted RHS to be hard to spot in the
rewritten graph.

Hopefully this can be made a little nicer in future updates!